requirements.txt
langchain==0.3.1
langgraph==0.2.34
langchain-openai==0.2.1
langchain-community==0.3.1
langchain-chroma==0.1.4
langchain-ollama==0.2.0
langchain-google-genai==2.0.0
langchain-groq==0.2.0
gradio==4.44.1
python-dotenv==1.0.1
bs4==0.0.2
wikipedia==1.4.0
pypdf==5.0.1
sentence-transformers==3.2.0
tavily-python==0.5.0
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7